refatoração da rota lote

+++ backend/models/lote.js
@@ -19,32 +19,34 @@ export async function getLote() {
   }
 }
 
-export async function getLoteAtividade(atividade) {
-  //testar para ver se funciona
-  try {
-    const lotes_ativos = await prisma.lote.findMany({ where: { atividade } })
-    return {
-      sucesso: true,
-      lotes_ativos,
-      message: "Lotes ativos listados com sucesso!!",
-    }
-  } catch (error) {
-    return {
-      sucesso: false,
-      message: "Erro ao listar lotes ativos!!",
-      error: error.message,
-    }
-  }
-}
+// N possui coluna atividade
+// export async function getLoteAtividade(atividade) {
+//   //testar para ver se funciona
+//   try {
+//     const lotes_ativos = await prisma.lote.findMany({ where: { atividade } })
+//     return {
+//       sucesso: true,
+//       lotes_ativos,
+//       message: "Lotes ativos listados com sucesso!!",
+//     }
+//   } catch (error) {
+//     return {
+//       sucesso: false,
+//       message: "Erro ao listar lotes ativos!!",
+//       error: error.message,
+//     }
+//   }
+// }
 
-export async function getLotePorAnimaliaId() {
+export async function getLotePorTipo(tipo) {
   try {
-    const lote_animalia = await prisma.lote.findMany({
-      where: { tipo: "Gado" || "Leite" },
-    })
+    const lote_tipo = await prisma.lote.findMany({
+      where: { tipo }
+    });
+
     return {
       sucesso: true,
-      lote_animalia,
+      lote_tipo,
       message: "Lotes de animalia listados com sucesso!!",
     }
   } catch (error) {
@@ -75,55 +77,45 @@ export async function getLoteRentabilidade(id_lote, rentabilidade) {
   }
 }
 
-export async function getLotePorDataColheita() {
-  try {
-    const lote_data_colheita = await prisma.lote.findMany({
-      where: { tipo: "Colheita" },
-    })
-    return {
-      sucesso: true,
-      lote_data_colheita,
-      message: "Lotes de colheita listados com sucesso!!",
-    }
-  } catch (error) {
-    return {
-      sucesso: false,
-      message: "Erro ao listar lotes de colheita!!",
-      error: error.message
-    }
-  } 
-}
-
-export async function getLotePorDataCriacao(ano, produtoId) { // testar
-  try {
-    const lote_data_criacao = await prisma.lote.findMany({
-      where: {
-        produtoId: Number(produtoId),   
-        dataFabricacao: {
-          gte: new Date(`${ano}-01-01`),
-          lte: new Date(`${ano}-12-31`)
-        },
-      },
-      select: {
-        nome: true,
-        tipo: true,
-        qntdItens: true,
-        produto: true
-      }
-    });
-    return {
-      sucesso: true,
-      lote_data_criacao,
-      message: "Lotes criados na data listados com sucesso!!",
-    }
-  } catch (error) {
-    return {
-      sucesso: false,
-      message: "Erro ao listar lotes criados na data!!",
-      error: error.message
-    }
-  }
-}
+// nao funciona pq n existe a coluna dataFabricacao, q é necessaria para listar os lotes criados
+// export async function getLotePorDataCriacao(ano, produtoId) { // testar
+//   try {
+//     const lote_data_criacao = await prisma.lote.findMany({
+//       where: {
+//         produtoId: Number(produtoId),   
+//         dataFabricacao: {
+//           gte: new Date(`${ano}-01-01`),
+//           lte: new Date(`${ano}-12-31`)
+//         },
+//       },
+//       select: {
+//         nome: true,
+//         tipo: true,
+//         qntdItens: true,
+//         preco: true,
+//         unidadeMedida: true,
+//         observacoes: true,
+//         statusQualidade: true,
+//         bloqueadoParaVenda: true,
+//         status: true,
+//         contratoId: true,
+//         dataEnvioReferencia: true,
+//         itensEsperados: true
+//       }
+//     });
+//     return {
+//       sucesso: true,
+//       lote_data_criacao,
+//       message: "Lotes criados na data listados com sucesso!!",
+//     }
+//   } catch (error) {
+//     return {
+//       sucesso: false,
+//       message: "Erro ao listar lotes criados na data!!",
+//       error: error.message
+//     }
+//   }
+// }
 
 export async function getlotePorId(id) {
   try {
@@ -142,20 +134,31 @@ export async function getlotePorId(id) {
   }
 }
 
-export async function createLote(data) {
+export async function createLote(data, unidadeId, contratoId) {
   try {
+    const unidade = await prisma.unidade.findUnique({where: { id: unidadeId }});
+    if (!unidade) {
+      return { sucesso: false, message: "Unidade não encontrada!" };
+    }
+
+    const contrato = await prisma.contrato.findUnique({where: { id: contratoId }});
+    if (!contrato) {
+      return { sucesso: false, message: "Contrato não encontrado!" };
+    }
+
+    const responsavel = await prisma.usuario.findUnique({ where: { id: data.responsavelId } });
+    if (!responsavel) {
+      return {sucesso: false, message: "Responsavel nao encontrado!!"}
+    }
+
     const novoLote = await prisma.lote.create({
       data: {
-        unidadeId: data.unidadeId,
-        responsavelId: data.responsavelId,
-        nome: data.nome,
-        tipo: data.tipo,
-        qntdItens: data.qntdItens,
-        observacoes: data.observacoes,
-        dataFabricacao: data.dataFabricacao,
-        dataValidade: data.dataValidade,
+        unidadeId: unidadeId,
+        contratoId: contratoId,
+        ...data
       },
-    })
+    });
+
     return {
       sucesso: true,
       novoLote,
@@ -170,10 +173,30 @@ export async function createLote(data) {
   }
 }
 
-export async function updateLote(id, data) {
+export async function updateLote(id, data, unidadeId, contratoId) {
   try {
+    const unidade = await prisma.unidade.findUnique({where: { id: unidadeId }});
+    if (!unidade) {
+      return { sucesso: false, message: "Unidade não encontrada!" };
+    }
+
+    const contrato = await prisma.contrato.findUnique({where: { id: contratoId }});
+    if (!contrato) {
+      return { sucesso: false, message: "Contrato não encontrado!" };
+    }
+
+    const responsavel = await prisma.usuario.findUnique({ where: { id: data.responsavelId } });
+    if (!responsavel) {
+      return {sucesso: false, message: "Responsavel nao encontrado!!"}
+    }
+
     const loteAtualizado = await prisma.lote.update({
       where: { id },
+      data: {
+        unidadeId: unidadeId,
+        contratoId: contratoId,
+        ...data
+      }
     })
     return {
       sucesso: true,
@@ -191,6 +214,11 @@ export async function updateLote(id, data) {
 
 export async function deleteLote(id) {
   try {
+    const lote = await prisma.lote.findUnique({ where: { id: id } });
+    if (!lote) {
+      return { sucesso: false, message: "Lote nao encontrado!" };
+    }
+
     const loteDeletado = await prisma.lote.delete({
       where: { id },
     })
